% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ehydro.R
\name{collect_ehydro}
\alias{collect_ehydro}
\title{collect_ehydro}
\usage{
collect_ehydro(ehydro_root, overwrite = FALSE, is_quiet = FALSE)
}
\arguments{
\item{ehydro_root}{path to location you want to store ehydro data}

\item{overwrite}{if TRUE will overwrite any processed ehydro tracker if FALSE will attempt to merge SurveyJob.shp with existing tracker, Default: FALSE}

\item{is_quiet}{is_quiet flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
collectes and formats ehydro SurveyJob.shp for use in the ePydRo workflow
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 ePydRo::collect_ehydro(ehydro_root = "~/data/ePydRo",overwrite=FALSE,is_quiet = FALSE)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[sf]{st_read}}, \code{\link[sf]{st_write}}
}
