% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marco.R
\name{marco}
\alias{marco}
\title{Hello world}
\usage{
marco()
}
\value{
a hello world message (polo!)
}
\description{
A hello world tester
}
\details{
Used to verify that ePydRo is successfully loaded.  Copy-paste-d from Google (https://ascii.co.uk/art/scuba).
}
\examples{
ePydRo::marco()
}
\seealso{
Other helper: 
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()}
}
\concept{helper}
