% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_error_block.R
\name{print_error_block}
\alias{print_error_block}
\title{prints error block}
\usage{
print_error_block()
}
\value{
print output
}
\description{
A error block helper
}
\details{
line art to draw eyes to printing notices.  Used when a catastrophic failure or critical logic error was encountered.
}
\examples{
ePydRo::print_error_block()
}
\seealso{
Other helper: 
\code{\link{marco}()},
\code{\link{print_warning_block}()}
}
\concept{helper}
