% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_vdatum.R, R/scrape_and_transform.R
\name{debug_vdatum}
\alias{debug_vdatum}
\alias{scrape_and_transform}
\title{debug_vdatum}
\usage{
debug_vdatum(ehydro_root, process_index, vdatum_domain_override)

scrape_and_transform(
  ehydro_root,
  process_index,
  overwrite = FALSE,
  vdatum_domain_override = NULL,
  is_quiet = FALSE
)
}
\arguments{
\item{ehydro_root}{path to location you want to store ehydro data}

\item{process_index}{index of sheet you want to stage}

\item{vdatum_domain_override}{passthrough for vdatum region override, can be ak, as, contiguous, gcnmi, prvi, westcoast, chesapeak_delaware, if left NULL will use 'contiguous', Default: NULL}

\item{overwrite}{overwrite files on disk if found, Default: FALSE}

\item{is_quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
TRUE

TRUE
}
\description{
debug_vdatum

download, normalize, and entwine ehydro
}
\details{
Puts string together to test the response of https://vdatum.noaa.gov/vdatumweb/

Stages data in ARCO form
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 ePydRo::debug_vdatum(ehydro_root, process_index = pr_surveys[3,]$sheet_index, vdatum_domain_override = 'prvi')
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 ePydRo::scrape_and_transform(ehydro_root = "~/data/ePydRo",process_index = 79,overwrite=FALSE,vdatum_domain_override=NULL,is_quiet = FALSE)
 }
}
}
\seealso{
\code{\link[sf]{st_read}}, \code{\link[sf]{st_layers}}, \code{\link[sf]{geos_unary}}, \code{\link[sf]{geos_combine}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_write}}
 \code{\link[httr]{GET}}, \code{\link[httr]{write_disk}}, \code{\link[httr]{http_error}}, \code{\link[httr]{content}}
 \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}
 \code{\link[glue]{glue}}
 \code{\link[lubridate]{decimal_date}}, \code{\link[lubridate]{ymd}}
 \code{\link[lidR]{LAS-class}}, \code{\link[lidR]{las_check}}, \code{\link[lidR]{writeLAS}}

\code{\link[sf]{st_read}}, \code{\link[sf]{st_layers}}, \code{\link[sf]{geos_unary}}, \code{\link[sf]{geos_combine}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_write}}
 \code{\link[httr]{GET}}, \code{\link[httr]{write_disk}}, \code{\link[httr]{http_error}}, \code{\link[httr]{content}}
 \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}
 \code{\link[glue]{glue}}
 \code{\link[lubridate]{decimal_date}}, \code{\link[lubridate]{ymd}}
 \code{\link[lidR]{LAS-class}}, \code{\link[lidR]{las_check}}, \code{\link[lidR]{writeLAS}}
}
