% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tiff_and_stage.R
\name{to_tiff_and_stage}
\alias{to_tiff_and_stage}
\title{to_tiff_and_stage}
\usage{
to_tiff_and_stage(ehydro_root, process_index, domain, is_quiet = FALSE)
}
\arguments{
\item{ehydro_root}{path to location you want to store ehydro data}

\item{process_index}{index of sheet you want to stage}

\item{domain}{PARAM_DESCRIPTION}

\item{is_quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates tiff from arco, dumps it out in a folder
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 to_tiff_and_stage(ehydro_root,process_index,domain,is_quiet = FALSE)
 }
}
}
\seealso{
\code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_zm}}
 \code{\link[glue]{glue}}
 \code{\link[data.table]{fread}}
 \code{\link[sfheaders]{sf_point}}
 \code{\link[raster]{rasterize}}, \code{\link[raster]{raster}}
 \code{\link[terra]{writeRaster}}
}
